noob_research_crew/pyproject.toml
[project]
name = "noob_research_crew"
version = "0.1.0"
description = "noob_research_crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "crewai[tools]>=0.86.0,<1.0.0",
    "firecrawl-py>=1.7.0",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.0",
]

[project.scripts]
noob_research_crew = "noob_research_crew.main:run"
run_crew = "noob_research_crew.main:run"
train = "noob_research_crew.main:train"
replay = "noob_research_crew.main:replay"
test = "noob_research_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
